test: use `librepo` for `centos-9` too
This commit switches centos-9 to use librepo as well. This is a
bit sad because it means we have no repo anymore that uses the
default "curl" backend. But even the centos9 mirrors are now so
unreliable that we cannot get things merged.

We should switch the default downloader here to librepo but we
need some downstream test that ensures that subscribed content
can still be installed with librepo (it should work but there
is no explicit test in our testsuite).